Require refactoring
Remove File.expand_path and use require_relative
Move matchers and helpers to support directory
Gem 'rubygems' is always required
