+++ lib/scanny.rb
@@ -1,8 +1,8 @@
-require File.expand_path(File.dirname(__FILE__) + "/scanny/issue")
-require File.expand_path(File.dirname(__FILE__) + "/scanny/report")
-require File.expand_path(File.dirname(__FILE__) + "/scanny/runner")
-require File.expand_path(File.dirname(__FILE__) + "/scanny/checks/check")
+require_relative "scanny/issue"
+require_relative "scanny/report"
+require_relative "scanny/runner"
+require_relative "scanny/checks/check"
 
-Dir[File.expand_path(File.dirname(__FILE__) + "/scanny/checks/**/*_check.rb")].each do |file|
+Dir[File.dirname(__FILE__) + "/scanny/checks/**/*_check.rb"].each do |file|
   require file
 end
