+++ spec/support/check_spec_helpers.rb
@@ -1,0 +1,5 @@
+module CheckSpecHelpers
+  def issue(*args)
+    Scanny::Issue.new("scanned_file.rb", 1, *args)
+  end
+end
